Move code from gui to guibase
